Связь продуктов и блюд, их удаление

+++ routes/add_product.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, request, redirect
-from flask_security import login_required, roles_accepted
-from datebase.classes import Product, db
+from flask_security import roles_accepted
+from datebase.classes import Product, AssociationDishProduct, Requisition, db
 from configs.app_configs import login_required
 
 
@@ -17,11 +17,11 @@ def add_product_page():
         measurement = request.form.get('measurement')
 
         if not name or not measurement:
-            return render_template('add_product.html')
+            return redirect('/add_product')
 
-        other_dish = db.session.query(Product).filter_by(name=name).first()
-        if other_dish:
-            return render_template('add_product.html')
+        other_product = db.session.query(Product).filter_by(name=name).first()
+        if other_product:
+            return redirect('/add_product')
 
         new_product = Product(name=name, measurement=measurement)
         db.session.add(new_product)
@@ -38,7 +38,15 @@ edit_product = Blueprint('edit_product', __name__, template_folder='templates')
 @roles_accepted('cook')
 def edit_product_page(id):
     product = db.session.query(Product).filter_by(id=id).first()
-    if request.method == 'POST':
+    if request.method == 'GET':
+        context = {
+            'name': product.name,
+            'measurement': product.measurement,
+        }
+        db.session.close()
+        return render_template('edit_product.html', **context)
+
+    elif request.method == 'POST':
         name = request.form.get('name')
         measurement = request.form.get('measurement')
 
@@ -57,13 +65,6 @@ def edit_product_page(id):
 
         return redirect('/read_product')
 
-    context = {
-        'name': product.name,
-        'measurement': product.measurement,
-    }
-    db.session.close()
-    return render_template('edit_product.html', **context)
-
 
 
 delete_product = Blueprint('delete_product', __name__, template_folder='templates')
@@ -72,7 +73,9 @@ delete_product = Blueprint('delete_product', __name__, template_folder='template
 @roles_accepted('cook')
 def delete_product_page(id):
     product = db.session.query(Product).filter_by(id=id).first()
-    if product.amount == 0:
+    dishes = db.session.query(AssociationDishProduct).filter_by(product_id=id).all()
+    requisitions = db.session.query(Requisition).filter_by(product_id=id).all()
+    if product.amount == 0 and not any([dishes, requisitions]):
         db.session.delete(product)
         db.session.commit()
     db.session.close()
